backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
groq==0.9.0
python-pptx==1.0.0
lxml==5.2.2
pdfplumber==0.11.0
slowapi==0.1.9
pydantic==2.7.0
